+++ experimental/navigation/README.md
@@ -1,4 +1,16 @@
-## OTel Instrumentation Library - React Native / Navigation
-  - expo-router
-  - react-native/nagivation
-  - react-native-navigation
+## OpenTelemetry Navigation Instrumentation for React Native
+Note: This is an experimental package under active development. New releases may include breaking changes.
+This module provides automatic instrumentation for
+  - [expo-router](https://docs.expo.dev/router/introduction/)
+  - [react-native/nagivation](https://reactnavigation.org/docs/getting-started)
+and
+  - [wix/react-native-navigation](https://wix.github.io/react-native-navigation/docs/before-you-start/)
+
+## Installation
+placeholder
+
+## Supported Versions
+placeholder
+
+## Usage
+placeholder
